fix: use bigserial and bigint in the database

+++ supabase/migrations/20241012171545_init_schema.sql
@@ -12,7 +12,7 @@ create type "public"."user_role" as enum('admin', 'manager', 'member');
 
 -- Tables
 create table if not exists "organizations" (
-	"id" serial primary key not null,
+	"id" bigserial primary key not null,
 	"name" text,
 	"domain" text not null,
 	"created_at" timestamp default now() not null,
@@ -21,7 +21,7 @@ create table if not exists "organizations" (
 
 create table if not exists "profiles" (
 	"id" uuid primary key not null references auth.users (id) on delete cascade on update cascade,
-	"organization_id" integer references organizations (id) on delete set default default 1 not null,
+	"organization_id" bigint references organizations (id) on delete set default default 1 not null,
 	"user_role" "user_role" default 'member' not null,
 	"created_at" timestamp default now() not null,
 	"updated_at" timestamp default now() not null
@@ -35,21 +35,21 @@ create table if not exists "role_permissions" (
 );
 
 create table if not exists "trainings" (
-	"id" serial primary key not null,
+	"id" bigserial primary key not null,
 	"title" text not null,
 	"tagline" text,
 	"description" text,
 	"image_url" text,
 	"preview_url" text,
 	"is_public" boolean default true not null,
-	"organization_id" integer references organizations (id) on delete cascade,
+	"organization_id" bigint references organizations (id) on delete cascade,
 	"created_at" timestamp default now() not null,
 	"updated_at" timestamp default now() not null
 );
 
 create table if not exists "modules" (
-	"id" serial primary key not null,
-	"training_id" integer not null references trainings (id) on delete cascade,
+	"id" bigserial primary key not null,
+	"training_id" bigint not null references trainings (id) on delete cascade,
 	"title" text not null,
 	"instructions" text,
 	"prompt" text,
@@ -61,15 +61,15 @@ create table if not exists "modules" (
 
 -- https://x.com/dshukertjr/status/1851231477671079952 Sensible default for user_id
 create table if not exists "enrollments" (
-	"id" serial primary key not null,
-	"training_id" integer not null references trainings (id) on delete cascade,
+	"id" bigserial primary key not null,
+	"training_id" bigint not null references trainings (id) on delete cascade,
 	"user_id" uuid default auth.uid() not null references profiles (id) on delete cascade,
 	"created_at" timestamp default now() not null
 );
 
 create table if not exists "history" (
-	"id" serial primary key not null,
-	"module_id" integer references modules (id) on delete set null,
+	"id" bigserial primary key not null,
+	"module_id" bigint references modules (id) on delete set null,
 	"user_id" uuid default auth.uid() not null references profiles (id) on delete cascade,
 	"assessment_text" text,
 	"assessment_score" integer,
